refactor: move message text rendering functionality to Message folder
