+++ src/index.ts
@@ -1,6 +1,4 @@
 export * from './components';
 export * from './context';
 export * from './i18n';
-// todo: distribute utils into separate files
 export * from './utils';
-export { getChannel } from './utils/getChannel';
