+++ src/utils/generateRandomId.ts
@@ -1,0 +1,6 @@
+import { nanoid } from 'nanoid';
+
+/**
+ * @deprecated will be removed in the next major release
+ */
+export const generateRandomId = nanoid;
