+++ src/utils/index.ts
@@ -1,0 +1,3 @@
+export * from './generateRandomId';
+export * from './getChannel';
+export * from './getWholeChar';
